refactor(clawdstrike): extract construct/receipts impl from engine.rs
Split two cohesive impl HushEngine blocks out of engine/mod.rs:
- construct.rs: constructors + builders (new/builder/with_policy/
  from_ruleset/with_*_guard/add_*_guard), HushEngineBuilder, Default, and
  build_custom_guards_from_policy. with_policy's guard-pipeline assembly
  moved verbatim.
- receipts.rs: create_receipt(_for_report) + signed variants + stats +
  reset + merge_report_metadata_into_receipt.

To let sibling impl blocks reach the engine's fields, HushEngine and
EngineState fields are widened private -> pub(crate) (internal only; the
public API is unchanged). HushEngineBuilder is re-exported from mod.rs so
crate::engine::HushEngineBuilder still resolves.

Pure relocation: no logic changed; 825 lib tests + formal-diff-tests pass.
